# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HomeBody Glam API",
        "version": "3.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "description": "User login",
                "operationId": "903626fe9a0cd6d42db6e883d556bc2e",
                "requestBody": {
                    "description": "User login info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_request",
                                    "password"
                                ],
                                "properties": {
                                    "login_request": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Passw0rd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Nguyen Van A"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "123456789"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "abcxyz1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Th  t b i, Vui l ng ki m tra l i t i kho n ho c m t kh u!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguy n V n A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nguyenvana@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirm": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Register success!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email invalid!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/partner/register": {
            "post": {
                "tags": [
                    "Partner"
                ],
                "summary": "Send a request to become a partner",
                "description": "Send a request to become a partner",
                "operationId": "54a17cdabff4aff4c5cf4e5a2d51a12b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the partner",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email of the partner",
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "Phone of the partner",
                                        "type": "string",
                                        "example": "123 456 7890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partner request sent successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/admin/partner-register/list": {
            "get": {
                "tags": [
                    "Admin Partner Register"
                ],
                "summary": "Get all partner registers",
                "description": "Get all partner registers",
                "operationId": "566a49914a59b344c09ac0110985738a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "123 456 7890"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2022-01-01 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/admin/partner-register/detail/{id}": {
            "get": {
                "tags": [
                    "Admin Partner Register"
                ],
                "summary": "Get partner detail by id",
                "description": "Get partner detail by id",
                "operationId": "getPartnerDetailById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of partner",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "123 456 7890"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2022-01-01 00:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partner not found"
                    }
                }
            }
        },
        "/api/admin/partner-register/update/{id}": {
            "put": {
                "tags": [
                    "Admin Partner Register"
                ],
                "summary": "Update a partner register status",
                "description": "Update a partner register status",
                "operationId": "updatePartnerRegisterStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of partner",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Partner register status to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Status of the partner register",
                                        "type": "string",
                                        "enum": [
                                            "APPROVED",
                                            "REJECTED",
                                            "DELETED"
                                        ],
                                        "example": "APPROVED"
                                    },
                                    "update": {
                                        "description": "Update the partner register status",
                                        "type": "string",
                                        "enum": [
                                            "Y",
                                            "N"
                                        ],
                                        "example": "Y"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partner register status updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Partner register not found"
                    }
                }
            }
        },
        "/api/admin/partner-register/delete/{id}": {
            "delete": {
                "tags": [
                    "Admin Partner Register"
                ],
                "summary": "Delete a partner register by ID",
                "description": "Delete a partner register by ID",
                "operationId": "98ea24b46c221a2a91587235f5773744",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of partner",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partner register deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Partner register not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to access this API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Partner",
            "description": "Partner"
        },
        {
            "name": "Admin Partner Register",
            "description": "Admin Partner Register"
        }
    ]
}